# Pathfinding performance improvements