… src/ClassicUO.Utility/ObjectPool.cs