- Primarily replaces the 15k length arrays and linear memory searches with a priority queue and dict.
- Includes a basic priority queue since .net 4 doesn't have one apparently. Implemented using a binary heap. Priority queue also includes a dict as well to implement O(1) `Contains`. 
- Been running for several hours now in a lumberjacking script, and so far so good.
- I'm seeing a 10x performance increase in cases where a path can be found (0.1 sec down to 0.01 sec). And seeing a worst-case performance improved 20x, where a path does not exist (0.4 to 0.6 sec down to 0.02 to 0.03 sec).

---
**Before:**
<img width="594" height="215" alt="image" src="https://github.com/user-attachments/assets/7bb26024-04d4-404b-bc20-ed2a896bba5e" />

---
**After:**
<img width="604" height="177" alt="image" src="https://github.com/user-attachments/assets/fa853943-b919-454a-84ea-1b505fd86b98" />


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **Refactor**
  * Improved pathfinding performance and efficiency by switching to dynamic data structures and a priority queue.
  * Increased the maximum path node capacity, allowing for more complex routes.
  * Enhanced reliability and maintainability of movement and auto-walk features.
  * Reduced memory allocations through object pooling and reusable collections.

* **New Features**
  * Pathfinding now supports longer and more intricate paths due to increased node limits.
  * Introduced a generic object pooling system to optimize resource management.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->